+++ interface/src/routes/connections/ntp/+page.svelte
@@ -1,10 +1,13 @@
 <script lang="ts">
-	import type { PageData } from './$types';
+	import type { PageData } from '../$types';
 	import NTP from './NTP.svelte';
 
 	export let data: PageData;
 </script>
 
-<div class="mx-8 my-8 flex flex-col gap-4">
+<div
+	class="mx-0 my-1 flex flex-col space-y-4
+     sm:mx-8 sm:my-8"
+>
 	<NTP />
 </div>