+++ interface/src/routes/system/status/+page.svelte
@@ -8,8 +8,8 @@
 </script>
 
 <div
-	class="mx-8 my-8 flex
-     flex-col space-y-4"
+	class="mx-0 my-1 flex flex-col space-y-4
+     sm:mx-8 sm:my-8"
 >
 	<SystemStatus />
 </div>