+++ interface/src/routes/wifi/+page.svelte
@@ -7,8 +7,8 @@
 </script>
 
 <div
-	class="mx-8 my-8 flex
-     flex-col gap-4"
+	class="mx-0 my-1 flex flex-col space-y-4
+     sm:mx-8 sm:my-8"
 >
 	<Wifi />
 	<Accesspoint />