+++ lib/framework/WiFiSettingsService.cpp
@@ -16,7 +16,8 @@
 
 WiFiSettingsService::WiFiSettingsService(AsyncWebServer *server, FS *fs, SecurityManager *securityManager, NotificationEvents *notificationEvents) : _httpEndpoint(WiFiSettings::read, WiFiSettings::update, this, server, WIFI_SETTINGS_SERVICE_PATH, securityManager),
                                                                                                                                                      _fsPersistence(WiFiSettings::read, WiFiSettings::update, this, fs, WIFI_SETTINGS_FILE),
-                                                                                                                                                     _lastConnectionAttempt(0)
+                                                                                                                                                     _lastConnectionAttempt(0),
+                                                                                                                                                     _notificationEvents(notificationEvents)
 {
     // We want the device to come up in opmode=0 (WIFI_OFF), when erasing the flash this is not the default.
     // If needed, we save opmode=0 before disabling persistence so the device boots with WiFi disabled in the future.
@@ -24,9 +25,6 @@ WiFiSettingsService::WiFiSettingsService(AsyncWebServer *server, FS *fs, Securit
     {
         WiFi.mode(WIFI_OFF);
     }
-
-    // Store SSE
-    _notificationEvents = notificationEvents;
 
     // Disable WiFi config persistance and auto reconnect
     WiFi.persistent(false);